--- src/main/java/DukeAbsentInfoException.java
@@ -1,11 +1,0 @@
-public class DukeAbsentInfoException extends DukeException {
-
-    DukeAbsentInfoException(String message) {
-        super(message);
-    }
-
-    @Override
-    public String toString() {
-        return this.message;
-    }
-}
