--- src/main/java/DukeIdxOOBException.java
@@ -1,8 +1,0 @@
-public class DukeIdxOOBException extends DukeException {
-    String message = "☹ OOPS!!! The index specified is out of bounds. ";
-
-    @Override
-    public String toString() {
-        return message;
-    }
-}
