+++ src/main/java/duke/DukeAbsentInfoException.java
@@ -1,0 +1,9 @@
+package duke;
+
+class DukeAbsentInfoException extends DukeException {
+
+    DukeAbsentInfoException(String message) {
+        super(message);
+    }
+
+}
