+++ src/main/java/duke/DukeException.java
@@ -1,9 +1,11 @@
-public class DukeException extends Exception {
+package duke;
+
+class DukeException extends Exception {
     
     String message;
 
     DukeException() {
-        this.message = "☹ OOPS!!! I'm sorry, but I don't know what that means :-(";
+        this.message = "OOPS!!! I'm sorry, but I don't know what that means :-(";
     }
 
     DukeException(String message) {
