+++ src/main/java/duke/DukeIdxOOBException.java
@@ -1,0 +1,9 @@
+package duke;
+
+class DukeIdxOOBException extends DukeException {
+
+    DukeIdxOOBException() {
+        super("☹ OOPS!!! The index specified is out of bounds.");
+    }
+
+}
