+++ src/main/java/duke/TaskList.java
@@ -1,6 +1,8 @@
+package duke;
+
 import java.util.ArrayList;
 
-public class TaskList {
+class TaskList {
     
     private ArrayList<Task> tasks;
